Modernize pipeline, restore license, add angle output and simple matcher
- Replace removed scipy.ndimage.imread/scipy.misc.imresize/imsave and
  deprecated np.int/np.float aliases so the pipeline runs on current
  numpy/scipy; fix skimage.measure.label connectivity arg and a
  float-vs-int bug in ridge_filter.py's meshgrid call.
- Refactor main_enhancement.py into a reusable extract_minutiae()
  function with an argparse CLI, path-independent I/O (no more
  hardcoded '../images/'), and --save-steps to dump intermediate
  pipeline images (normalized/orientation/filtered/skeleton) for
  reports/presentations.
- Compute a ridge-angle for every minutia (from the orientation field)
  and include it in minutiae.txt and the returned minutiae list.
- Add match_fingerprints.py: a simplified minutiae-based fingerprint
  comparison CLI, for demonstrating the matching concept.
- Restore the BSD-2-Clause LICENSE (the code derives from
  Utkarsh-Deshmukh/Fingerprint-Enhancement-Python).
- Add requirements.txt.
- Stop tracking compiled .pyc/__pycache__ files and the runtime-
  generated src/data.txt; remove the unused, buggy ridge_freq_2.py
  duplicate; .gitignore updated accordingly.
- Rewrite README with setup/usage instructions for both CLIs.

+++ src/ridge_filter.py
@@ -50,7 +50,7 @@ Created on Fri Apr 22 03:15:03 2016
 
 
 import numpy as np
-import scipy;
+import scipy.ndimage;
 def ridge_filter(im, orient, freq, kx, ky):
     angleInc = 3;
     im = np.double(im);
@@ -77,8 +77,8 @@ def ridge_filter(im, orient, freq, kx, ky):
     sigmax = 1/unfreq[0]*kx;
     sigmay = 1/unfreq[0]*ky;
     
-    sze = np.round(3*np.max([sigmax,sigmay]));
-    
+    sze = int(np.round(3*np.max([sigmax,sigmay])));
+
     x,y = np.meshgrid(np.linspace(-sze,sze,(2*sze + 1)),np.linspace(-sze,sze,(2*sze + 1)));
     
     reffilter = np.exp(-(( (np.power(x,2))/(sigmax*sigmax) + (np.power(y,2))/(sigmay*sigmay)))) * np.cos(2*np.pi*unfreq[0]*x); # this is the original gabor filter
